+++ src/pharmpy/modeling/datainfo.py
@@ -2,8 +2,112 @@ from collections.abc import Mapping
 from pathlib import Path
 from typing import Any, Union, overload
 
+from pharmpy.deps import pandas as pd
 from pharmpy.internals.fs.path import normalize_user_given_path, path_absolute
-from pharmpy.model import DataInfo, DataVariable, Model
+from pharmpy.model import ColumnInfo, DataInfo, DataVariable, Model
+
+
+def _read_dataset_header_and_separator(path) -> tuple[list[str], str]:
+    with open(path) as file:
+        first_line = file.readline()
+        if ',' not in first_line:
+            colnames = list(pd.read_csv(path, nrows=0, sep=r'\s+'))
+            separator = r'\s+'
+        else:
+            colnames = list(pd.read_csv(path, nrows=0))
+            separator = ','
+    if len(colnames) > 0:
+        colnames[0] = colnames[0].lstrip('#')
+    return colnames, separator
+
+
+def create_datainfo(path_or_df: Union[str, Path, pd.DataFrame]) -> DataInfo:
+    """Create a DataInfo
+
+    Will assume NONMEM names of columns
+
+    Parameters
+    ----------
+    path_or_df : Path | str | pd.DataFrame
+        A path to a dataset or a dataset
+
+    Returns
+    -------
+    DataInfo
+        DataInfo object
+    """
+
+    if not isinstance(path_or_df, pd.DataFrame):
+        path = normalize_user_given_path(path_or_df)
+        path = path_absolute(path)
+        datainfo_path = path.with_suffix('.datainfo')
+        try:
+            di = read_datainfo(datainfo_path)
+        except FileNotFoundError:
+            pass
+        else:
+            di = di.replace(path=path)
+            return di
+
+        colnames, separator = _read_dataset_header_and_separator(path)
+
+    else:
+        colnames = path_or_df.columns
+        separator = ","
+        path = None
+
+    column_info = []
+    for colname in colnames:
+        colname = colname.replace('.', '_')  # pandas uses . to name mangle
+        if colname == 'ID' or colname == 'L1':
+            var = DataVariable.create(colname, type='id', scale='nominal')
+            info = ColumnInfo.create(colname, var, datatype='int32')
+        elif colname == 'DV':
+            var = DataVariable.create(colname, type='dv')
+            info = ColumnInfo.create(colname, var)
+        elif colname == 'TIME':
+            if not set(colnames).isdisjoint({'DATE', 'DAT1', 'DAT2', 'DAT3'}):
+                datatype = 'nmtran-time'
+            else:
+                datatype = 'float64'
+            var = DataVariable.create(colname, type='idv', scale='ratio')
+            info = ColumnInfo.create(colname, var, datatype=datatype)
+        elif colname == 'EVID':
+            var = DataVariable.create(colname, type='event', scale='nominal')
+            info = ColumnInfo.create(colname, var)
+        elif colname == 'MDV':
+            if 'EVID' in colnames:
+                var = DataVariable.create(colname, type='mdv')
+                info = ColumnInfo.create(colname, var)
+            else:
+                var = DataVariable.create(colname, type='event', scale='nominal')
+                info = ColumnInfo.create(colname, var, datatype='int32')
+        elif colname == 'AMT':
+            var = DataVariable.create(colname, type='dose', scale='ratio')
+            info = ColumnInfo.create(colname, var)
+        elif colname == 'RATE':
+            var = DataVariable.create(colname, type='rate', scale='ratio')
+            info = ColumnInfo.create(colname, var)
+        elif colname == 'BLQ':
+            var = DataVariable.create(colname, type='blq', scale='nominal')
+            info = ColumnInfo.create(colname, var, datatype='int32')
+        elif colname == 'LLOQ':
+            var = DataVariable.create(colname, type='lloq', scale='ratio')
+            info = ColumnInfo.create(colname, var)
+        elif colname == 'DVID':
+            var = DataVariable.create(colname, type='dvid', scale='nominal')
+            info = ColumnInfo.create(colname, var, datatype='int32')
+        elif colname == 'SS':
+            var = DataVariable.create(colname, type='ss', scale='nominal')
+            info = ColumnInfo.create(colname, var, datatype='int32')
+        elif colname == 'II':
+            var = DataVariable.create(colname, type='ii', scale='ratio')
+            info = ColumnInfo.create(colname, var)
+        else:
+            info = ColumnInfo.create(colname)
+        column_info.append(info)
+    di = DataInfo.create(column_info, path=path, separator=separator)
+    return di
 
 
 def read_datainfo(path: Union[str, Path]) -> DataInfo: